make bls validator its own module and re-export it
