+++ dpp/src/dash_platform_protocol.rs
@@ -1,7 +1,8 @@
 use std::sync::Arc;
+use crate::BlsValidator;
 
 use crate::errors::DashPlatformProtocolInitError;
-use crate::identity::validation::{BlsValidator, NativeBlsValidator, PublicKeysValidator};
+use crate::identity::validation::{PublicKeysValidator};
 use crate::identity::IdentityFacade;
 use crate::version::{ProtocolVersionValidator, COMPATIBILITY_MAP, LATEST_VERSION};
