+++ dpp/src/tests/fixtures/get_public_keys_validator_fixture.rs
@@ -1,4 +1,5 @@
-use crate::identity::validation::{PublicKeysValidator, PUBLIC_KEY_SCHEMA, PUBLIC_KEY_SCHEMA_FOR_TRANSITION, NativeBlsValidator};
+use crate::identity::validation::{PublicKeysValidator, PUBLIC_KEY_SCHEMA, PUBLIC_KEY_SCHEMA_FOR_TRANSITION};
+use crate::NativeBlsValidator;
 
 pub fn get_public_keys_validator_for_transition() -> PublicKeysValidator<NativeBlsValidator> {
     PublicKeysValidator::new_with_schema(PUBLIC_KEY_SCHEMA_FOR_TRANSITION.clone(), NativeBlsValidator::default()).unwrap()
