+++ dpp/src/tests/identity/validation/public_keys_validator_spec.rs
@@ -1,8 +1,8 @@
 use serde_json::{json, Value};
 
-use crate::assert_consensus_errors;
+use crate::{assert_consensus_errors, NativeBlsValidator};
 use crate::consensus::ConsensusError;
-use crate::identity::validation::{NativeBlsValidator, PublicKeysValidator};
+use crate::identity::validation::{PublicKeysValidator};
 use crate::identity::validation::TPublicKeysValidator;
 use crate::identity::{KeyType, Purpose, SecurityLevel};
 use crate::tests::fixtures::get_public_keys_validator;
